Added a warning that we should be using OrmRecordNotFound.
When getting record, we are not raising an exception if not found. Instead
we are simply returning nil. We should set a standard. Either we return
nil on no records found or we use OrmRecordNotFound but not both.

As a side note, if keeping OrmRecordNotFound it should have Exception
added to the end of the name.